+++ packages/client/src/v2-events/components/forms/inputs/FileInput/DocumentUploaderWithOption.tsx
@@ -254,7 +254,7 @@ function DocumentUploaderWithOption({
             value={selectedOption}
             onChange={(val) => {
               void helpers.setTouched(true)
-              setSelectedOption(val)
+              setSelectedOption(val ?? undefined)
               setUnselectedOptionError('')
             }}
           />